split java bin scripts to allow separate build/test/pacakge

+++ bin/smi/package.py
@@ -27,7 +27,9 @@ def runtime_platform() -> str:
 
 def main(argv: Optional[Sequence[str]] = None) -> int:
 
-    parser = C.get_parser()
+    parser = argparse.ArgumentParser()
+    C.add_clean_arg(parser)
+    C.add_tag_arg(parser)
     DC.add_args(parser)
     parser.add_argument(
         "--no-build",
@@ -37,7 +39,7 @@ def main(argv: Optional[Sequence[str]] = None) -> int:
 
     dist_tag_dir = C.DIST_DIR / args.tag
 
-    if args.clean:
+    if args.clean and dist_tag_dir.is_dir():
         shutil.rmtree(dist_tag_dir)
     dist_tag_dir.mkdir(parents=True, exist_ok=True)
